+++ Tests/TestApps/DialogHelperDemo/App.xaml.cs
@@ -1,11 +1,32 @@
-﻿using System.Windows;
+﻿#region Header
+// --------------------------------------------------------------------------
+// Tethys.Silverlight
+// ==========================================================================
+//
+// This library contains common code for WPF, Silverlight, Windows Phone and
+// Windows 8 projects.
+//
+// ===========================================================================
+//
+// <copyright file="App.xaml.cs" company="Tethys">
+// Copyright  2010-2015 by Thomas Graf
+//            All rights reserved.
+//            Licensed under the Apache License, Version 2.0.
+//            Unless required by applicable law or agreed to in writing, 
+//            software distributed under the License is distributed on an
+//            "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+//            either express or implied. 
+// </copyright>
+//
+// ---------------------------------------------------------------------------
+#endregion
 
 namespace WpfLayoutDemo
 {
   /// <summary>
-  /// Interaction logic for App.xaml
+  /// Interaction logic for App.
   /// </summary>
-  public partial class App : Application
+  public partial class App
   {
-  }
-}
+  } // App
+} // WpfLayoutDemo