+++ Tests/TestApps/DialogHelperDemo/MessageDialogWin7.xaml.cs
@@ -1,15 +1,39 @@
-﻿using System.Windows;
+﻿#region Header
+// --------------------------------------------------------------------------
+// Tethys.Silverlight
+// ==========================================================================
+//
+// This library contains common code for WPF, Silverlight, Windows Phone and
+// Windows 8 projects.
+//
+// ===========================================================================
+//
+// <copyright file="MessageDialogWin7.xaml.cs" company="Tethys">
+// Copyright  2010-2015 by Thomas Graf
+//            All rights reserved.
+//            Licensed under the Apache License, Version 2.0.
+//            Unless required by applicable law or agreed to in writing, 
+//            software distributed under the License is distributed on an
+//            "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+//            either express or implied. 
+// </copyright>
+//
+// ---------------------------------------------------------------------------
+#endregion
 
 namespace WpfLayoutDemo
 {
-  /// <summary>
-  /// Interaction logic for SimpleDialog.xaml
-  /// </summary>
-  public partial class MessageDialogWin7 : Window
-  {
-    public MessageDialogWin7()
+    /// <summary>
+    /// Interaction logic for MessageDialogWin7.
+    /// </summary>
+    public partial class MessageDialogWin7
     {
-      InitializeComponent();
-    }
-  }
-}
+        /// <summary>
+        /// Initializes a new instance of the <see cref="MessageDialogWin7"/> class.
+        /// </summary>
+        public MessageDialogWin7()
+        {
+            InitializeComponent();
+        } // MessageDialogWin7()
+    } // MessageDialogWin7
+} // WpfLayoutDemo