+++ Tests/TestApps/WpfApplicationExceptionDialog/App.xaml.cs
@@ -1,53 +1,79 @@
-﻿using System;
-using System.Windows;
-using System.Windows.Threading;
-using TgLib.Diagnostics;
+﻿#region Header
+// --------------------------------------------------------------------------
+// Tethys.Silverlight
+// ==========================================================================
+//
+// This library contains common code for WPF, Silverlight, Windows Phone and
+// Windows 8 projects.
+//
+// ===========================================================================
+//
+// <copyright file="App.xaml.cs" company="Tethys">
+// Copyright  2010-2015 by Thomas Graf
+//            All rights reserved.
+//            Licensed under the Apache License, Version 2.0.
+//            Unless required by applicable law or agreed to in writing, 
+//            software distributed under the License is distributed on an
+//            "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+//            either express or implied. 
+// </copyright>
+//
+// ---------------------------------------------------------------------------
+#endregion
 
 namespace WpfApplicationExceptionDialog
 {
-  /// <summary>
-  /// Interaction logic for App.xaml
-  /// </summary>
-  public partial class App : Application
-  {
-    public App()
-    {
-      DispatcherUnhandledException += OnApplicationException;
-      AppDomain.CurrentDomain.UnhandledException += OnAppDomainException;
-    } // App()
+    using System;
+    using System.Windows.Threading;
+
+    using Tethys.Silverlight.Diagnostics;
 
     /// <summary>
-    /// Called when an application exception is thrown.
+    /// Interaction logic for App.
     /// </summary>
-    /// <param name="sender">The sender.</param>
-    /// <param name="e">The
-    ///  <see cref="System.Windows.Threading.DispatcherUnhandledExceptionEventArgs"/> 
-    /// instance containing the event data.</param>
-    static void OnApplicationException(object sender, 
-      DispatcherUnhandledExceptionEventArgs e)
+    public partial class App
     {
-      AppErrorWindow win = new AppErrorWindow(e.Exception);
-      win.Owner = Current.MainWindow;
-      win.ShowDialog();
-      e.Handled = true;
-    } // OnApplicationException()
+        /// <summary>
+        /// Initializes a new instance of the <see cref="App"/> class.
+        /// </summary>
+        public App()
+        {
+            DispatcherUnhandledException += OnApplicationException;
+            AppDomain.CurrentDomain.UnhandledException += OnAppDomainException;
+        } // App()
 
-    /// <summary>
-    /// Called when an AppDomain exception is thrown.
-    /// </summary>
-    /// <param name="sender">The sender.</param>
-    /// <param name="e">The <see cref="System.UnhandledExceptionEventArgs"/>
-    ///  instance containing the event data.</param>
-    static void OnAppDomainException(object sender, 
-      UnhandledExceptionEventArgs e)
-    {
-      Exception ex = e.ExceptionObject as Exception;
-      if (ex != null)
-      {
-        AppErrorWindow win = new AppErrorWindow(ex);
-        win.Owner = Current.MainWindow;
-        win.ShowDialog();
-      } // if
-    } // OnAppDomainException()
-  } // App
+        /// <summary>
+        /// Called when an application exception is thrown.
+        /// </summary>
+        /// <param name="sender">The sender.</param>
+        /// <param name="e">The
+        ///  <see cref="System.Windows.Threading.DispatcherUnhandledExceptionEventArgs"/> 
+        /// instance containing the event data.</param>
+        private static void OnApplicationException(object sender,
+          DispatcherUnhandledExceptionEventArgs e)
+        {
+            var win = new AppErrorWindow(e.Exception);
+            win.Owner = Current.MainWindow;
+            win.ShowDialog();
+            e.Handled = true;
+        } // OnApplicationException()
+
+        /// <summary>
+        /// Called when an AppDomain exception is thrown.
+        /// </summary>
+        /// <param name="sender">The sender.</param>
+        /// <param name="e">The <see cref="System.UnhandledExceptionEventArgs"/>
+        /// instance containing the event data.</param>
+        private static void OnAppDomainException(object sender,
+          UnhandledExceptionEventArgs e)
+        {
+            var ex = e.ExceptionObject as Exception;
+            if (ex != null)
+            {
+                AppErrorWindow win = new AppErrorWindow(ex);
+                win.Owner = Current.MainWindow;
+                win.ShowDialog();
+            } // if
+        } // OnAppDomainException()
+    } // App
 } // WpfApplicationExceptionDialog