+++ Tests/TestApps/WpfApplicationExceptionDialog/MainWindow.xaml.cs
@@ -1,21 +1,52 @@
-﻿using System.Windows;
+﻿#region Header
+// --------------------------------------------------------------------------
+// Tethys.Silverlight
+// ==========================================================================
+//
+// This library contains common code for WPF, Silverlight, Windows Phone and
+// Windows 8 projects.
+//
+// ===========================================================================
+//
+// <copyright file="MainWindow.xaml.cs" company="Tethys">
+// Copyright  2010-2015 by Thomas Graf
+//            All rights reserved.
+//            Licensed under the Apache License, Version 2.0.
+//            Unless required by applicable law or agreed to in writing, 
+//            software distributed under the License is distributed on an
+//            "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+//            either express or implied. 
+// </copyright>
+//
+// ---------------------------------------------------------------------------
+#endregion
 
 namespace WpfApplicationExceptionDialog
 {
-  /// <summary>
-  /// Interaction logic for MainWindow.xaml
-  /// </summary>
-  public partial class MainWindow : Window
-  {
-    public MainWindow()
-    {
-      InitializeComponent();
-    }
+    using System.Windows;
 
-    private void BtnThrowException(object sender, RoutedEventArgs e)
+    /// <summary>
+    /// Interaction logic for MainWindow.
+    /// </summary>
+    public partial class MainWindow
     {
-      int x = 0;
-      int a = 1/x;
-    }
-  }
-}
+        /// <summary>
+        /// Initializes a new instance of the <see cref="MainWindow"/> class.
+        /// </summary>
+        public MainWindow()
+        {
+            InitializeComponent();
+        } // MainWindow()
+
+        /// <summary>
+        /// Handles a click on the the <c>BtnThrowException</c> control.
+        /// </summary>
+        /// <param name="sender">The sender.</param>
+        /// <param name="e">The <see cref="RoutedEventArgs"/> instance containing the event data.</param>
+        private void BtnThrowException(object sender, RoutedEventArgs e)
+        {
+            int x = 0;
+            int a = 1 / x;
+        }
+    } // MainWindow
+} // WpfApplicationExceptionDialog