+++ ChangeLog.md
@@ -1,6 +1,13 @@
 Tethys.Silverlight Change Log
 =============================
 
+##V1.2.1	(2015-04-03)##
+* Windows Phone 7.1 support cancelled.
+* Existing test application in solution included.
+* There is a new static class DesignTime to hold
+  the property `IsInDesignMode`. The old implemntation
+  in `ViewModelBase` has been marked obsolete.
+
 ##V1.2.0	(2015-01-04)##
 * Moved to Github
 * Many additional classes added.